Move cell into file
